requirements.txt
beautifulsoup4
PyPDF2
python-docx